move interface to /models
