+++ src/models/Student.ts
@@ -9,3 +9,15 @@ export interface Student {
   msg: string;
   feedback?: string;
 }
+
+export interface IRawStudent {
+  firstName: string;
+  lastName: string;
+  email: string;
+  subjects: string;
+  msg?: string;
+  verified: boolean;
+  alreadyScreened: boolean;
+  phone?: string;
+  birthday?: Date;
+}
